Graphics.Displayable

Documentation

data MouseEvent

Constructors

MouseDown

MouseUp

MouseMove

class Displayable t where

The type class of things that can be displayed.

For the click function, the first point is the location of the click, and the second point is the upper-left corner of the thing being clicked.

The measure function gives a fixed dimension for all values of a type, appropriate for table formatting. The flowMeasure function gives a specific measurement for each value, appropriate for a flow layout.

Methods

measure :: t -> (Int32, Int32)

flowMeasure :: t -> (Int32, Int32)

draw :: HDC -> (Int32, Int32) -> t -> IO ()

mouse :: MouseEvent -> HWND -> IORef [HWND] -> (Int32, Int32) -> (Int32, Int32) -> t -> IO ()

newtype Str

Constructors

Str String

newtype Legend

A legend for a graph.

Constructors

Legend [String]

data Bars

A set of bars that compare related data points side by side. A list of these is a bar chart.

Constructors

Bars

Fields

values :: ![Double]

maxValue :: !Double

newtype Plot

A scatter plot. The outer list elements are for different datasets that are drawn with different colours.

Constructors

Plot [[(Double, Double)]]

data Selection

A fixed list of values the user can select from with a popup menu.

Constructors

Selection

Fields

selection :: !(IORef Int32)

choices :: ![String]

data Clear

When used in a flow layout, moves subsequent text below all floated elements.

Constructors

Clear

data Disp

A type box that forgets everything about a value except the fact that it is displayable.

Constructors

forall t . Displayable t => Disp t

newtype HorizontalList t

A list that displays horizontally instead of vertically.

Constructors

HorizontalList [t]

attachDisplay :: Displayable t => IORef t -> HWND -> IO ()

Subclasses the window in order to display the value in the reference.

onChange :: Displayable t => IORef t -> HWND -> IO ()

Call this when the reference changes.

createNiceFont :: IO HFONT