Graphics.Layout

Documentation

data Just

Constructors

L

Center

R

Justified

data Floated

Constructors

Inline

FloatLeft

FloatRight

type Paragraph t = (Just, [(t, Floated)])

width :: Displayable t => t -> Int32

height :: Displayable t => t -> Int32

layout :: Displayable t => Int32 -> ((Int32, Int32), (Int32, Int32), Int32) -> [Paragraph t] -> [(Int32, Int32, (t, Floated))]

hitTest :: (Bounded b, Num b, Ord b) => (b, b) -> [(b, b, t)] -> Maybe (b, b, t)

attachLayout :: Displayable t => IORef [Paragraph t] -> HWND -> IO (IO ())

Call this to attach flow layout to the given window. Returns a procedure that should be called when the reference changes.